[CL-3174] Delete more dead code
