+++ front/app/utils/projectUtils.ts
@@ -1,15 +1,13 @@
 import {
+  IIdeaJsonFormSchemas,
   CustomFieldCodes,
-  IIdeaFormSchemas,
-} from 'services/ideaCustomFieldsSchemas';
-import { IIdeaJsonFormSchemas } from 'services/ideaJsonFormsSchema';
+} from 'services/ideaJsonFormsSchema';
 import { Locale } from 'typings';
 import { isNilOrError } from './helperUtils';
 
 export function isFieldEnabled(
   fieldCode: CustomFieldCodes,
   ideaCustomFieldsSchemas:
-    | IIdeaFormSchemas['data']['attributes']
     | IIdeaJsonFormSchemas['data']['attributes']
     | undefined
     | null
